backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0